mintlib: sync atomic types with kernel and improve lazy initialization
- Update pthread_once_t, sem_t, and spinlock types to 'long' to match updated kernel structures.
- Add pthread_setup_threading_np() checks to pthread primitives (cancel, cleanup, mutex, cond, etc.) to ensure threading is initialized before syscalls.
- Simplify pthread_once logic to rely directly on kernel CAS.
- Remove redundant manual count check in sem_trywait to avoid race conditions.
- Fix __sleep to correctly handle the main thread vs child threads in multithreaded mode.
- Remove commented-out dead code in pthread_spin.c.

+++ include/semaphore.h
@@ -23,9 +23,9 @@ extern "C" {
 typedef struct
 {
     struct thread *wait_queue;      /* Queue of threads waiting on this sem */
-    volatile short count;           /* Current semaphore count */
+    volatile long count;           /* Current semaphore count */
     /* Non threaded values */
-    volatile short io_count;        /* Reference count for named sems */
+    volatile long io_count;        /* Reference count for named sems */
     char sem_id[SEM_NAME_MAX + 1];  /* Fixed array instead of pointer */
 } sem_t;
